Move application menu bar code to a new file
Put the menu code in its own file to keep it more organised. We're also
going to be adding a lot more code here for the base main menu.

Wasn't sure if I should be releasing these menu objects, but it looks
like the application works when I do release them, so maybe
`setMainMenu` does a retain or copy.

I can't decide if I should use C-style identifiers or Foundation-style
in the menu file.