MainMenu: Trying to get Edit>Find menu items working
I learned about tags and that there are new names for the NSTextFinder
identifiers starting in 10.7. But these still don't seem to be doing
anything.

But I tried adding an NSTextView to a window in a base Xcode project and
the Find actions don't seem to work there either. So now that I have
this, I'm thinking the rest of the problem lies elsewhere. Perhaps the
NSTextView isn't connected enough to enable the Find actions.