MainMenu: Try switching to `NSFontManager` -fontMenu:
This creates the whole font menu for me, and the actions work correctly,
very nice! Read about this in the `NSFontManager` documentation.

The only hitch is that some menu items don't appear to have the proper
keyEquivalents set. For example, "Show Colors", "Copy Style", and "Paste
Style" are all missing the Apple key from the shortcuts.