MainMenu: Add missing Command modifier if missing from Font menu items
I don't understand why, but on Mac OS X 10.15, three menu items in the
automatically-generated Font menu don't have a Command key modifier.
Detect when this happens and add the modifier to the mask.

These menu items correctly had a Command key modifier when I tested the
code on macOS 13.