Makefile: Try to build .app bundle
I'm trying to set up the Make targets to build a .app bundle, but I'm
having trouble handling file names with spaces.

I sort of managed to do it using the strategy articulated by andrewdotn
(https://stackoverflow.com/users/14558/andrewdotn) with "${@}" in this
Stack Overflow answer:
https://stackoverflow.com/questions/14639906/can-gnu-make-handle-spaces/14640047#14640047

However, it doesn't seem to be working in the `subst` or `patsubst`
calls using "%" for the localisation files. I get the error:

    make: *** No rule to make target `en.lproj', needed by `app'.  Stop.

The error looks like it's saying that the
`build/$(APP_NAME).app/Contents/Resources/%.lproj` rule couldn't be
found, even though it is declared.

It looks like I'm going to have to explore other options to handle file
names, or at least application names, with spaces.

I copied the Info.plist file from Mass-menu and updated some fields to
work with this project. I also copied and modified the Make rules from
Mass-menu, but that project doesn't need to handle spaces in file names.