Makefile: Test replacing spaces in the app name with a differen char
My idea is to use the no-space version of the app name in the Make
targets, and mv the files to the with-spaces version when everything is
built.

Decided to use the Unicode ZERO WIDTH NO-BREAK SPACE character U+FEFF as
the sentinel replacement character. We can always change it to something
else (like a string of unique characters) later if needed.

I took the Perl command from penguin359
(https://unix.stackexchange.com/users/6167/penguin359) on Stack
Exchange:
https://unix.stackexchange.com/questions/12273/in-bash-how-can-i-convert-a-unicode-codepoint-0-9a-f-into-a-printable-charact/12279#12279