Makefile: Move space-less version to a .app bundle with spaces
This gets us the dependencies set up so they can use paths without
spaces, but creates a .app bundle with spaces as a final product.

It does so by moving the files where we substituted spaces with a
non-space character to file names with spaces.

I don't like this yet because `make app` builds a whole new binary and
everything else, but the resulting bundle is what we want.