Makefile: Try to not recompile the executable when building .app
Keep the temporary .app bundle with substituted spaces around and copy
the bundle directory so we don't end up rebuilding the executable file
when `make` is re-executed.

This does fail on `make` re-execution because the space-substituted
binary file was moved and no longer exists. Might have to think about
how to make that cleaner later.

The more pressing concern is that the final .app bundle with spaces in
it reports an error when I try to open it:

    $ open build/Base\ Windowed\ Application.app
    The application cannot be opened because its executable is missing.

Not sure what the problem is there yet.