Info.plist: Change executable name to "Base Windowed Application"
This is why I was getting the "The application cannot be opened because
its executable is missing." error.

Now the application opens correctly from the .app bundle.

However, this revealed another problem with the make recipes: The
Info.plist file isn't copied from the space-substituted bundle directory
to the final .app directory.