Makefile: Adjust no-space rules so we don't rebuild every run
Now, `make app` doesn't error and says "Nothing to be done" when there
are no changes.

I adjusted the recipes to make the final bundle with spaces dependent on
all files in the no-space bundle.

Switch to `rsync` from `cp` so only the files that did change are
copied.

When updating the MacOS binary file, don't move it, otherwise that
triggers a recompile. Instead copy it to the with-spaces .app bundle. We
also need to remove the no-spaces version of the executable from the
with-spaces .app bundle.

Now I need to work out how to make all this work on application names
that don't have spaces in them.