Makefile: Remove errors when APP_NAME has no spaces
Add some conditionals to remove the targets that rename the bundle the
application name has spaces in it. This fixes the warnings and circular
dependency problems we had when building an application bundle with no
spaces where `APP_NAME` and `APP_NAME_NOSPACE` are the same.

The rules and organisation definitely need to be cleaned up, but the
idea works. We should eliminate one of these conditions and put all the
related rules together under a single `ifneq`.