Make a document-based application
Add the basic necessities for a document-based application. This is what
I came up with on 2023-09-10. The `DocumentWindowController` was just an
initial idea and doesn't do anything: we create the required window
controller in `Document`.

The `CFBundleDocumentTypes` entry enables us to interact with text
files.

This confirms that the "Open Recent" menu does get added automatically
below the "Open" menu.