Document: Initial attempt to cascade new document windows
When creating a new document with Apple-N, try to cascade the new window
so that it opens a bit below and to the right of the previous window.

Store the previous window origin in a static variable for a relative
reference point.

Currently this seems to cascade, but not by a full titlebar height and
corresponding horizontal position. Needs some adjustment.